delete comma in verse 8
This change follows the suggestion of Carol Brinneman, one of the "readers" assigned to check the ULB. Tom Warren passed this on to me to implement (or not), and since the change in programs from etherpad to this, I will try this out to see how this works. Any correction or advice would be much appreciated.